+++ common/src/main/java/net/xstopho/resource_backpacks/backpack/BackpackItem.java
@@ -7,7 +7,6 @@ import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.SimpleMenuProvider;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.BlockItem;
@@ -15,7 +14,6 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.ItemContainerContents;
 import net.minecraft.world.item.context.UseOnContext;
-import net.minecraft.world.item.equipment.Equippable;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.xstopho.resource_backpacks.backpack.tooltip.CompactTooltipComponent;
@@ -35,17 +33,18 @@ public class BackpackItem extends BlockItem {
     private final BackpackLevel backpackLevel;
 
     public BackpackItem(Block block, BackpackLevel backpackLevel, Properties properties) {
-        super(block, properties.useBlockDescriptionPrefix()
-                .component(DataComponents.CONTAINER, ItemContainerContents.EMPTY)
-                .component(DataComponents.EQUIPPABLE, Equippable.builder(EquipmentSlot.CHEST).build()));
+        super(block, properties.component(DataComponents.CONTAINER, ItemContainerContents.EMPTY));
         this.backpackLevel = backpackLevel;
     }
 
+
+
     @Override
-    public InteractionResult use(Level level, Player player, InteractionHand hand) {
+    public InteractionResult use(Level level, Player player, InteractionHand usedHand) {
+        ItemStack stack = player.getItemInHand(usedHand);
 
         if (!level.isClientSide && BackpackConfig.openFromInventory) {
-            player.openMenu(getMenuProvider(player.getItemInHand(hand)));
+            player.openMenu(getMenuProvider(stack));
         }
 
         return InteractionResult.SUCCESS;
@@ -55,7 +54,7 @@ public class BackpackItem extends BlockItem {
     public InteractionResult useOn(UseOnContext context) {
         Player player = context.getPlayer();
 
-        if (player != null && player.isCrouching()) {
+        if (player != null && (player.isCrouching() || !BackpackConfig.openFromInventory)) {
             super.useOn(context);
         }