+++ common/src/main/java/net/xstopho/resource_backpacks/mixin/common/ShapedRecipeMixin.java
@@ -9,6 +9,7 @@ import net.minecraft.world.item.crafting.CraftingInput;
 import net.minecraft.world.item.crafting.ShapedRecipe;
 import net.xstopho.resource_backpacks.backpack.BackpackItem;
 import net.xstopho.resource_backpacks.backpack.util.BackpackLevel;
+import net.xstopho.resource_backpacks.mixin.accessor.ShapedRecipeAccessor;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;