+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-#### v0.11.6-BETA
-- fixing quick move functionality
+#### v0.12.0-BETA
+- adding Backpack slot for Survival Inventory
+    - remove the ability to equip the Backpack on the Chest slot
+- adding proper BlockStates when Backpack is placed on a wall
+- fixing Inventory Label Color for Backpacks, based on used style
+- internal network changes
+    - on forge this produces log spam on the client, when showing Ender Backpack Tooltip